# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Outlets Service",
    "description": "Manages press outlets (publications) and their domain authority data",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "components": {
    "schemas": {
...
      "CreateCategory": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "categoryName": {
            "type": "string",
            "minLength": 1
          },
          "scope": {
            "type": "string",
            "enum": [
              "city",
              "state_or_province",
              "country",
              "multi-country_region",
              "international"
            ]
          },
          "region": {
            "type": "string"
          },
          "exampleOutlets": {
            "type": "string"
          },
          "whyRelevant": {
            "type": "string",
            "default": ""
          },
          "whyNotRelevant": {
            "type": "string",
            "default": ""
          },
          "relevanceScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "default": 0
          }
        },
        "required": [
          "campaignId",
          "categoryName"
        ]
      },
      "UpdateCategory": {
        "type": "object",
        "properties": {
          "categoryName": {
            "type": "string",
            "minLength": 1
          },
          "scope": {
            "type": "string",
            "enum": [
              "city",
              "state_or_province",
              "country",
              "multi-country_region",
              "international"
            ]
          },
          "region": {
            "type": "string",
            "nullable": true
          },
          "exampleOutlets": {
            "type": "string",
            "nullable": true
          },
          "whyRelevant": {
            "type": "string"
          },
          "whyNotRelevant": {
            "type": "string"
          },
          "relevanceScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authenticating requests"
      }
    }
  },
  "paths": {
...
    "/categories": {
      "post": {
        "summary": "Create press category",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created"
          }
        }
      },
      "get": {
        "summary": "List categories by campaign",
        "parameters": [
          {
            "in": "query",
            "name": "campaignId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category list"
          }
        }
      }
    },
    "/categories/{id}": {
      "patch": {
        "summary": "Update category",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/categories/{id}/status": {
      "patch": {
        "summary": "Update category status",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category status updated"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/internal/outlets/by-ids": {
      "get": {
        "summary": "Batch lookup outlets by IDs",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated outlet IDs"
          }
        ],
        "responses": {
          "200": {
            "description": "Outlets found"
          }
        }
      }
    },
    "/internal/outlets/by-campaign/{campaignId}": {
      "get": {
        "summary": "All outlets for a campaign with DR data",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign outlets with DR"
          }
        }
      }
    },
...
  }
}